requirements.txt
pandas
sqlalchemy
psycopg2-binary
langchain
langchain-community
langchain-openai
langchain-groq
python-dotenv
